src/frontend/requirements.txt
python-dotenv==1.2.2
streamlit==1.55.0
requests==2.32.3